refactor SO into its own service. add urban dictionary define.
