+++ bot.js
@@ -1,8 +1,8 @@
 var constants = require('./constants')
+var services = require('./services')
 
 // Loads environment files
 require('envc')({})
-var request = require('request')
 
 if (!process.env.token) {
   console.log('Error: Specify token in environment')
@@ -16,96 +16,12 @@ var controller = Botkit.slackbot({
   debug: true
 })
 
-var bot = controller.spawn({
+controller.spawn({
   token: process.env.token
 }).startRTM()
 
-function logError (err, msg) {
-  if (err) {
-    var message = msg || 'Danger Will Robinson!'
-    bot.botkit.log(message, err)
-  }
-}
-
-controller.hears([constants.LOOKUP], constants.ADDRESSED, function (bot, message) {
-  var text = message.text
-  // Verify first word is lookup
-  var firstWord = text.substr(0, text.indexOf(' '))
-  if (firstWord === constants.LOOKUP) {
-    var search = text.substr(text.indexOf(' ') + 1)
-    var questionQS = {
-      order: 'desc',
-      q: search,
-      site: constants.API.stack.site,
-      sort: 'relevance'
-    }
-
-    request.get({
-      url: `${constants.API.stack.url}/search/advanced`,
-      gzip: true,
-      qs: questionQS,
-      headers: {
-        'Content-Type': 'application/json; charset=utf-8'
-      }
-    }, (err, res, body) => {
-      if (err) {
-        logError('Error With Stack Response', err)
-        bot.reply('Error: You must construct additional pylons!')
-        return
-      }
-      var results = JSON.parse(body)
-
-      // Answer exists
-      if (results.items.length) {
-        var question = results.items.find((item) => item.is_answered && item.accepted_answer_id)
-        if (!question) {
-          bot.reply('No answered result found!')
-          return
-        }
-        var resultQuestion = question.title
-        var answerQS = {
-          filter: 'withbody',
-          order: 'desc',
-          site: constants.API.stack.site,
-          sort: 'activity'
-        }
-        request.get({
-          url: `${constants.API.stack.url}/answers/${question.accepted_answer_id}`,
-          gzip: true,
-          qs: answerQS,
-          headers: {
-            'Content-Type': 'application/json; charset=utf-8'
-          }
-        }, (err, res, body) => {
-          if (err) {
-            logError('Error With Stack Response', err)
-            bot.reply('Error: You must construct additional pylons!')
-            return
-          }
-          var answers = JSON.parse(body)
-
-          if (answers.items.length) {
-            var answer = answers.items[0]
-            var pieces = answer.body.split(constants.REGEX.pre_code)
-            var stripped = pieces.map((piece) => {
-              if (constants.REGEX.pre_code.test(piece)) {
-                var codeBlock = piece.replace(/<pre><code>/igm, '```')
-                codeBlock = codeBlock.replace(/<\/code><\/pre>/igm, '```')
-                return codeBlock
-              }
-              var textBlock = piece.replace(/<\/*code>/igm, '`')
-              textBlock = textBlock.replace(/(<([^>]+)>)/igm, '')
-              return textBlock
-            }).join('')
-
-            bot.reply(message, `*Q:* \`${resultQuestion}\`\n`)
-            bot.reply(message, `*A:* ${stripped}`)
-          }
-        })
-      }
-    })
-  }
-})
+controller.hears([constants.LOOKUP], constants.ADDRESSED, services.stackOverflow)
+controller.hears([constants.DEFINE], constants.ADDRESSED, services.urbanDictionary)
 
 controller.hears(['uptime'], constants.ADDRESSED, function (bot, message) {
   var hostname = os.hostname()
