+++ services/index.js
@@ -1,0 +1,7 @@
+var stackOverflow = require('./stack-overflow')
+var urbanDictionary = require('./urban-dictionary')
+
+module.exports = {
+  stackOverflow,
+  urbanDictionary
+}
